requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
supabase==2.11.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
jinja2==3.1.4
email-validator==2.2.0
httpx==0.28.1
